+++ tools/bun/src/proto.rs
@@ -10,7 +10,6 @@ extern "ExtismHost" {
 }
 
 static NAME: &str = "Bun";
-static BIN: &str = "bun";
 
 #[plugin_fn]
 pub fn register_tool(Json(_): Json<ToolMetadataInput>) -> FnResult<Json<ToolMetadataOutput>> {
@@ -18,7 +17,8 @@ pub fn register_tool(Json(_): Json<ToolMetadataInput>) -> FnResult<Json<ToolMeta
         name: NAME.into(),
         type_of: PluginType::Language,
         config_schema: Some(SchemaBuilder::build_root::<BunPluginConfig>()),
-        plugin_version: Some(env!("CARGO_PKG_VERSION").into()),
+        minimum_proto_version: Some(Version::new(0, 42, 0)),
+        plugin_version: Version::parse(env!("CARGO_PKG_VERSION")).ok(),
         self_upgrade_commands: vec!["upgrade".into()],
         ..ToolMetadataOutput::default()
     }))
@@ -73,7 +73,7 @@ pub fn download_prebuilt(
 
     let mut avx2_suffix = "";
 
-    if env.arch == HostArch::X64 && env.os == HostOS::Linux && command_exists(&env, "grep") {
+    if env.arch == HostArch::X64 && env.os.is_linux() && command_exists(&env, "grep") {
         let output = exec_command!("grep", ["avx2", "/proc/cpuinfo"]);
 
         if output.exit_code != 0 {
@@ -125,7 +125,7 @@ pub fn locate_executables(
         // `bunx` isn't a real binary provided by Bun so we can't symlink it.
         // Instead, it's simply the `bun` binary named `bunx` and Bun toggles
         // functionality based on `args[0]`.
-        exe_link_path: Some(env.os.get_exe_name(BIN).into()),
+        exe_link_path: Some(env.os.get_exe_name("bun").into()),
 
         // The approach doesn't work for shims since we use child processes,
         // so execute `bun x` instead (notice the space).
@@ -135,12 +135,14 @@ pub fn locate_executables(
     };
 
     Ok(Json(LocateExecutablesOutput {
-        globals_lookup_dirs: vec!["$HOME/.bun/bin".into()],
-        primary: Some(ExecutableConfig::new(env.os.get_exe_name(BIN))),
-        secondary: HashMap::from_iter([
-            // bunx
+        exes: HashMap::from_iter([
+            (
+                "bun".into(),
+                ExecutableConfig::new_primary(env.os.get_exe_name("bun")),
+            ),
             ("bunx".into(), bunx),
         ]),
+        globals_lookup_dirs: vec!["$HOME/.bun/bin".into()],
         ..LocateExecutablesOutput::default()
     }))
 }