+++ tools/go/tests/shims_test.rs
@@ -1,4 +1,4 @@
 use proto_pdk_test_utils::*;
 
 #[cfg(not(windows))]
-generate_shims_test!("go-test");
+generate_shims_test!("go-test", ["go"]);