+++ tools/internal-schema/tests/shims_test.rs
@@ -4,6 +4,6 @@ use starbase_sandbox::locate_fixture;
 #[cfg(not(windows))]
 generate_shims_test!(
     "schema-test",
-    [],
+    ["schema-test"],
     Some(locate_fixture("schemas").join("base.toml"))
 );