+++ tools/node-depman/tests/shims_test.rs
@@ -2,19 +2,19 @@
 mod npm {
     use proto_pdk_test_utils::*;
 
-    generate_shims_test!("npm-test", ["npx", "node-gyp"]);
+    generate_shims_test!("npm-test", ["npm", "npx", "node-gyp"]);
 }
 
 #[cfg(not(windows))]
 mod pnpm {
     use proto_pdk_test_utils::*;
 
-    generate_shims_test!("pnpm-test", ["pnpx"]);
+    generate_shims_test!("pnpm-test", ["pnpm", "pnpx"]);
 }
 
 #[cfg(not(windows))]
 mod yarn {
     use proto_pdk_test_utils::*;
 
-    generate_shims_test!("yarn-test", ["yarnpkg"]);
+    generate_shims_test!("yarn-test", ["yarn", "yarnpkg"]);
 }