+++ tools/node/tests/shims_test.rs
@@ -2,4 +2,4 @@
 use proto_pdk_test_utils::*;
 
 #[cfg(not(windows))]
-generate_shims_test!("node-test");
+generate_shims_test!("node-test", ["node"]);