+++ tools/proto/src/proto.rs
@@ -12,7 +12,8 @@ pub fn register_tool(Json(_): Json<ToolMetadataInput>) -> FnResult<Json<ToolMeta
     Ok(Json(ToolMetadataOutput {
         name: "proto".into(),
         type_of: PluginType::CommandLine,
-        plugin_version: Some(env!("CARGO_PKG_VERSION").into()),
+        minimum_proto_version: Some(Version::new(0, 42, 0)),
+        plugin_version: Version::parse(env!("CARGO_PKG_VERSION")).ok(),
         self_upgrade_commands: vec!["up".into(), "upgrade".into()],
         ..ToolMetadataOutput::default()
     }))
@@ -88,7 +89,7 @@ pub fn locate_executables(
 ) -> FnResult<Json<LocateExecutablesOutput>> {
     let env = get_host_environment()?;
 
-    let mut primary = ExecutableConfig::new(env.os.get_exe_name("proto"));
+    let mut primary = ExecutableConfig::new_primary(env.os.get_exe_name("proto"));
     primary.no_bin = true;
     primary.no_shim = true;
 
@@ -97,8 +98,7 @@ pub fn locate_executables(
     secondary.no_shim = true;
 
     Ok(Json(LocateExecutablesOutput {
-        primary: Some(primary),
-        secondary: HashMap::from_iter([("proto-shim".into(), secondary)]),
+        exes: HashMap::from_iter([("proto".into(), primary), ("proto-shim".into(), secondary)]),
         ..LocateExecutablesOutput::default()
     }))
 }