+++ tools/rust/src/proto.rs
@@ -2,6 +2,7 @@ use crate::helpers::*;
 use crate::toolchain_toml::ToolchainToml;
 use extism_pdk::*;
 use proto_pdk::*;
+use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
 
@@ -29,7 +30,8 @@ pub fn register_tool(Json(_): Json<ToolMetadataInput>) -> FnResult<Json<ToolMeta
             override_dir: Some(get_toolchain_dir(&env)?),
             version_suffix: Some(format!("-{}", get_target_triple(&env, NAME)?)),
         },
-        plugin_version: Some(env!("CARGO_PKG_VERSION").into()),
+        minimum_proto_version: Some(Version::new(0, 42, 0)),
+        plugin_version: Version::parse(env!("CARGO_PKG_VERSION")).ok(),
         ..ToolMetadataOutput::default()
     }))
 }
@@ -213,11 +215,12 @@ pub fn locate_executables(
     // Binaries are provided by Cargo (`~/.cargo/bin`), so don't create
     // our own shim and bin. But we do need to ensure that the install
     // worked, so we still check for the `cargo` binary.
-    let mut primary = ExecutableConfig::new(env.os.get_exe_name("bin/cargo"));
+    let mut primary = ExecutableConfig::new_primary(env.os.get_exe_name("bin/cargo"));
     primary.no_bin = true;
     primary.no_shim = true;
 
     Ok(Json(LocateExecutablesOutput {
+        exes: HashMap::from_iter([("cargo".into(), primary)]),
         exes_dir: Some("bin".into()),
         globals_lookup_dirs: vec![
             "$CARGO_INSTALL_ROOT/bin".into(),
@@ -225,7 +228,6 @@ pub fn locate_executables(
             "$HOME/.cargo/bin".into(),
         ],
         globals_prefix: Some("cargo-".into()),
-        primary: Some(primary),
         ..LocateExecutablesOutput::default()
     }))
 }