+++ tools/internal-schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+#### 💥 Breaking
+
+- Removed `platform.*.bin_path` support (use `install.exes` instead).
+- Deprecated `install.primary` and `install.secondary` (use `install.exes` instead).
+
+#### 🚀 Updates
+
+- Added `install.exes` for defining primary and secondary executables.
+- Updated to support proto v0.42 release.
+
 ## 0.15.1
 
 #### 🚀 Updates